AI Implemented with the same framework as the Player
Units take turns based on their speed via the turn manager class.

The Turn manager waits for input from the current unit. Upon receiving input the turnmanger will then invoke the IUnitTurnAction that was found.
Unit controllers have the methods that will find the action to be performed on the current unit's turn. In the player's case, this will be determined by button presses. The AI will need to have logic to find what action and direction they want.