Made small editors for unit profiles and modelobjects
There are window tools that can be used for easier creation of unit assets.
There is also a small preview of what the model will look like.
This preview is created using unity Editor class where it calls on the onPreviewGUI functions. This creates a playbleGraph somewhere which must also include an equal destroy playable graph when not in use. But the editor class has no access to the creation or destroy which only becomes a problem after the player has been tested. When the player is stopped and the state returns back to unity editor, somewhere along the way Playable graphs get cleaned up and the editor is not updated with this information properly, it loses it's reference without being destroyed and leaks the data. To avoid this, I have chosen to destroy the editor when the player is started and create a new one upon re entering edit mode. The preview does not work during runtime tests rather than leaking the memory.