Overhauled the editor for unitProfiles
The editor makes use of a SerializedObject targeting the unitprofile, and the fields are drawn dynamically by finding the SerializedProperty by it's name. This should make the process of extending the fields easier later on.

Added many and various helper functions that can be called from the static class MysteryEditorUtility