Completely changed the unit stats system
The new system is dynamic so that the user can add and remove stats easily whenever they want.

Units use stats via a dictionary using statKeys as the key.
There are 2 enums that are accessed.
CoreStatKey
ResourceStatKey
These can be freely changed by the user, but the profiles will need to be "cleaned" as serialisation of these stats will have missing references or not exist yet.

The profiles can be edited by the profiles editor.

Core Stats are simply floats that do not change during gameplay. eg. speed is 5, strength is 4.
Resource Stats are float with a current value and a max value. eg. health is 75/100, magic points is 50/50, experience points is 23/200.
In the profile editor, only the max value is needed to be recorded. The user can set the unit's current value during runtime however they wish.

Models are still loaded and referenced by an older method and I would like to update it.