Inventory system is implemented
Inventory and items are generic to enable targeting a specified target object that would "hold" the item.
In this case this will be the Unit class.
Currently there is only examples of consumable items. These items call a functions "Use" in order to change the value of a Resource stat. eg. a health potion is used and the unit's current health is increased.