Implemented a token class
The GameMapToken is the class that controls how any form of token is placed or moved around the board.
The unit class derives from it as units need to be moveable tokens.

ItemToken also derives from it, but needs to be considered a different type from Units.
Currently Items and Units can inhabit the same tile. This let's units "pick up" items on the ground.