Needed To Release objects in Object Pools
Objects weren't being released from the pool, so the objectPool thought that all objects were active. Objects are now "Released" when the object should be set In active.

Updated UI Functionality for in game menus. Test buttons are "I" for inventory, "O" for settings placeholder. and "ESC" for open last panel.