Refactoring code
The main culprit here is that In order for "Characters" to be saved. Some form of serialisation needs to be present. Unity doesn't serialise Dictionarys. So I created a Serialisable Dictionary class that uses a List to keep track of keys and values. But this meant that in order to use it, the unitstats needed to use them. So I decided to clean up the code as I went and replaced the old dictionarys. This would break a large amount of the code base and the rditor scripts also have become obsolete.